feat: add shared NPC sprite animator component
